[#9] refactor: Account 패키지 조절
